updated the test cases and handled boolean in numberType
